779: Implement Consent Response JWT signing in RCS (#86)
* 779: Implement Consent Response JWT signing in RCS

RCS was calling out to IG to do JWT signing, which is not secure. In this change RCS will now sign its own JWTs.

Added RcsJwtSigner class which implements signing. This requires that a signingKeyId, signingAlgorithm and privateKeyPath are provided via spring config.

Added test-jwt-signing-key.pem which is an RSA private key which can be used for integration testing purposes.

Fixed a bug in the way that the iss claim was being set in the consent response JWT. The old logic was copying the iss claim from the request JWT, which is AM. Now we plug the value in from spring config, it must match the Remote Consent configuration id in AM

Issue: https://github.com/SecureBankingAccessToolkit/SecureBankingAccessToolkit/issues/779