239: Configuring RCS to use internal IG svc to access consent repo
The IG Consent Repo (/repo route), is an internal feature. Use the internal IG address to access it, so that the ingress rule can be removed.

Note: In future the /repo route will be removed in favour of a dedicated consent API

Cleaned up config to remove jwks related item which are no longer used. Removed JWT validation logic as this is now being done in IG.

Issue: https://github.com/SecureBankingAccessToolkit/SecureBankingAccessToolkit/issues/239