915: Renaming FilePaymentConsentDetails initiation field to filePayment
The RCS UI codebase is expecting a filePayment field in the details response. Reverting the initiation field name back to filePayment.

The UI code has a single object modelling all consent details, therefore we cannot use the initiation field for file data as it does not match the initiation schema of the other payments.

https://github.com/SecureApiGateway/SecureApiGateway/issues/915