1023: Changing Consent Entity definition to store the FR data model representation of a Consent Request
Decision has been made to store the FR data model representation because OBIE introduce breaking changes often in their minor API upgrades. Using the FR data model allows us to lessen the impact of some of these issues by applying workarounds in the conversion layer.

For payments, we also store the charges (which get added to the consent response) as the FR data model representation.

Removing the requestType field from the Entity definition, as it is now redundant.

https://github.com/SecureApiGateway/SecureApiGateway/issues/1023