1047: Consents which are long-lived support re-authentication
Re-authentication means that a PSU can Authorise (or Reject) a consent that is currently in an Authorised state. This is only supported for long-lived consents, which are currently: Account Access and Domestic VRP consents.

Updating the ConsentStateModel to add support for re-authentication. Adding a new VRPConsentStateModel, and using this for VRP Consents (rather than the PaymentConsentStateModel).

As part of modelling the VRPConsentStateModel, we are also fixing a bug which meant that VRP consents could be Consumed. This state is not supported for VRP payments as they are long-lived (all other payments are short-lived)

Adding new ConsentStoreException error: CONSENT_REAUTHENTICATION_NOT_SUPPORTED, which will be raised if we attempt to fetch the consent details for the UI for a consent that has already been authorised (and re-authentication is not supported).

https://github.com/SecureApiGateway/SecureApiGateway/issues/1047